+++ src/java/level2/newsclustering/Main.java
@@ -1,0 +1,9 @@
+package java.level2.newsclustering;
+
+public class Main {
+
+    public static void main(String[] args) {
+
+    }
+}
+
