+++ src/java/level2/numberconversion/Main.java
@@ -1,0 +1,22 @@
+package java.level2.numberconversion;
+
+public class Main {
+
+    public static void main(String[] args) {
+
+        Solution solution = new Solution();
+        int x = 10;
+        int y = 40;
+        int n = 5;
+        int answer = solution.solution(x, y, n);
+
+        System.out.println("answer = " + answer);
+    }
+}
+
+class Solution {
+    public int solution(int x, int y, int n) {
+        int answer = 0;
+        return answer;
+    }
+}
