--- src/level2/newsclustering/Main.java
@@ -1,9 +1,0 @@
-package level2.newsclustering;
-
-public class Main {
-
-    public static void main(String[] args) {
-
-    }
-}
-
